add Kingfisher for image loading and cashing.
